Convert `streamlit hello` to a MPA
