+++ lib/streamlit/cli.py
@@ -149,10 +149,10 @@ def main_docs():
 @configurator_options
 def main_hello(**kwargs):
     """Runs the Hello World script."""
-    from streamlit.hello import hello
+    from streamlit.hello import Hello
 
     bootstrap.load_config_options(flag_options=kwargs)
-    filename = hello.__file__
+    filename = Hello.__file__
     _main_run(filename, flag_options=kwargs)
